requirements.txt
beautifulsoup4==4.9.1
blessed==1.17.6
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
idna==2.10
inquirer==2.7.0
python-editor==1.0.4
python-slugify==4.0.1
readchar==2.0.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
tabulate==0.8.7
text-unidecode==1.3
urllib3==1.25.10
wcwidth==0.2.5
